made a bunch of parsers private
